# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BGarden API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Biometry/specimen/{specimenId}": {
      "get": {
        "tags": [
          "Biometry"
        ],
        "parameters": [
          {
            "name": "specimenId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BiometryDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BiometryDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BiometryDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Phenology/specimen/{specimenId}": {
      "get": {
        "tags": [
          "Phenology"
        ],
        "parameters": [
          {
            "name": "specimenId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PhenologyDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PhenologyDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PhenologyDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Specimen/sector/{sectorType}": {
      "get": {
        "tags": [
          "Specimen"
        ],
        "parameters": [
          {
            "name": "sectorType",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SectorType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpecimenDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpecimenDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpecimenDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Specimen/filter": {
      "get": {
        "tags": [
          "Specimen"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "familyId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "regionId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpecimenDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpecimenDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpecimenDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Specimen/{id}": {
      "get": {
        "tags": [
          "Specimen"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SpecimenDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecimenDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecimenDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Specimen"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpecimenDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SpecimenDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SpecimenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SpecimenDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecimenDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecimenDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Specimen"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Specimen": {
      "post": {
        "tags": [
          "Specimen"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpecimenDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SpecimenDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SpecimenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SpecimenDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecimenDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecimenDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Specimen/with-images": {
      "post": {
        "tags": [
          "Specimen"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Id": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "InventoryNumber": {
                    "type": "string"
                  },
                  "SectorType": {
                    "$ref": "#/components/schemas/SectorType"
                  },
                  "Latitude": {
                    "type": "number",
                    "format": "double"
                  },
                  "Longitude": {
                    "type": "number",
                    "format": "double"
                  },
                  "LocationWkt": {
                    "type": "string"
                  },
                  "RegionId": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "RegionName": {
                    "type": "string"
                  },
                  "FamilyId": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "FamilyName": {
                    "type": "string"
                  },
                  "RussianName": {
                    "type": "string"
                  },
                  "LatinName": {
                    "type": "string"
                  },
                  "Genus": {
                    "type": "string"
                  },
                  "Species": {
                    "type": "string"
                  },
                  "Cultivar": {
                    "type": "string"
                  },
                  "Form": {
                    "type": "string"
                  },
                  "Synonyms": {
                    "type": "string"
                  },
                  "DeterminedBy": {
                    "type": "string"
                  },
                  "PlantingYear": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "SampleOrigin": {
                    "type": "string"
                  },
                  "NaturalRange": {
                    "type": "string"
                  },
                  "EcologyAndBiology": {
                    "type": "string"
                  },
                  "EconomicUse": {
                    "type": "string"
                  },
                  "ConservationStatus": {
                    "type": "string"
                  },
                  "ExpositionId": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "ExpositionName": {
                    "type": "string"
                  },
                  "HasHerbarium": {
                    "type": "boolean"
                  },
                  "DuplicatesInfo": {
                    "type": "string"
                  },
                  "OriginalBreeder": {
                    "type": "string"
                  },
                  "OriginalYear": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "Country": {
                    "type": "string"
                  },
                  "Illustration": {
                    "type": "string"
                  },
                  "Notes": {
                    "type": "string"
                  },
                  "FilledBy": {
                    "type": "string"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              },
              "encoding": {
                "Id": {
                  "style": "form"
                },
                "InventoryNumber": {
                  "style": "form"
                },
                "SectorType": {
                  "style": "form"
                },
                "Latitude": {
                  "style": "form"
                },
                "Longitude": {
                  "style": "form"
                },
                "LocationWkt": {
                  "style": "form"
                },
                "RegionId": {
                  "style": "form"
                },
                "RegionName": {
                  "style": "form"
                },
                "FamilyId": {
                  "style": "form"
                },
                "FamilyName": {
                  "style": "form"
                },
                "RussianName": {
                  "style": "form"
                },
                "LatinName": {
                  "style": "form"
                },
                "Genus": {
                  "style": "form"
                },
                "Species": {
                  "style": "form"
                },
                "Cultivar": {
                  "style": "form"
                },
                "Form": {
                  "style": "form"
                },
                "Synonyms": {
                  "style": "form"
                },
                "DeterminedBy": {
                  "style": "form"
                },
                "PlantingYear": {
                  "style": "form"
                },
                "SampleOrigin": {
                  "style": "form"
                },
                "NaturalRange": {
                  "style": "form"
                },
                "EcologyAndBiology": {
                  "style": "form"
                },
                "EconomicUse": {
                  "style": "form"
                },
                "ConservationStatus": {
                  "style": "form"
                },
                "ExpositionId": {
                  "style": "form"
                },
                "ExpositionName": {
                  "style": "form"
                },
                "HasHerbarium": {
                  "style": "form"
                },
                "DuplicatesInfo": {
                  "style": "form"
                },
                "OriginalBreeder": {
                  "style": "form"
                },
                "OriginalYear": {
                  "style": "form"
                },
                "Country": {
                  "style": "form"
                },
                "Illustration": {
                  "style": "form"
                },
                "Notes": {
                  "style": "form"
                },
                "FilledBy": {
                  "style": "form"
                },
                "images": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SpecimenDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecimenDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecimenDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BiometryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "specimenId": {
            "type": "integer",
            "format": "int32"
          },
          "specimenInfo": {
            "type": "string",
            "nullable": true
          },
          "measurementDate": {
            "type": "string",
            "format": "date-time"
          },
          "height": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "flowerDiameter": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PhenologyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "specimenId": {
            "type": "integer",
            "format": "int32"
          },
          "specimenInfo": {
            "type": "string",
            "nullable": true
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "floweringStart": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "floweringEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "fruitingDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SectorType": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "SpecimenDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "inventoryNumber": {
            "type": "string",
            "nullable": true
          },
          "sectorType": {
            "$ref": "#/components/schemas/SectorType"
          },
          "latitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "locationWkt": {
            "type": "string",
            "nullable": true
          },
          "regionId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "regionName": {
            "type": "string",
            "nullable": true
          },
          "familyId": {
            "type": "integer",
            "format": "int32"
          },
          "familyName": {
            "type": "string",
            "nullable": true
          },
          "russianName": {
            "type": "string",
            "nullable": true
          },
          "latinName": {
            "type": "string",
            "nullable": true
          },
          "genus": {
            "type": "string",
            "nullable": true
          },
          "species": {
            "type": "string",
            "nullable": true
          },
          "cultivar": {
            "type": "string",
            "nullable": true
          },
          "form": {
            "type": "string",
            "nullable": true
          },
          "synonyms": {
            "type": "string",
            "nullable": true
          },
          "determinedBy": {
            "type": "string",
            "nullable": true
          },
          "plantingYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "sampleOrigin": {
            "type": "string",
            "nullable": true
          },
          "naturalRange": {
            "type": "string",
            "nullable": true
          },
          "ecologyAndBiology": {
            "type": "string",
            "nullable": true
          },
          "economicUse": {
            "type": "string",
            "nullable": true
          },
          "conservationStatus": {
            "type": "string",
            "nullable": true
          },
          "expositionId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "expositionName": {
            "type": "string",
            "nullable": true
          },
          "hasHerbarium": {
            "type": "boolean"
          },
          "duplicatesInfo": {
            "type": "string",
            "nullable": true
          },
          "originalBreeder": {
            "type": "string",
            "nullable": true
          },
          "originalYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "illustration": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "filledBy": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}